requirements.txt
streamlit
python-dotenv
requests
#huggingface_hub